refactor: use subparsers + generically handle API data after validating seasons
