+++ src/objekt_stats/__init__.py
@@ -2,7 +2,10 @@ import argparse
 import asyncio
 import sys
 
-from objekt_stats.enums import Artist, Edition, Season, StatsOutput
+from objekt_stats.enums.artms import ArtmsSeason
+from objekt_stats.enums.cli import StatsOutput
+from objekt_stats.enums.cosmo import Artist, Edition, Season
+from objekt_stats.enums.triples import TripleSSeason
 from objekt_stats.objekts.service import default_objekt_service
 from objekt_stats.signals import register_signal_handlers
 
@@ -17,42 +20,51 @@ class CosmoStatsArgsNamespace:
     output: StatsOutput
 
 
+ARTIST_SUBPARSERS = frozenset(
+    [
+        (Artist.TRIPLES, TripleSSeason),
+        (Artist.ARTMS, ArtmsSeason),
+    ]
+)
+
+
 def main() -> None:
     args = CosmoStatsArgsNamespace()
 
     parser = argparse.ArgumentParser(prog="objekt-stats")
-    parser.add_argument(
-        "artist",
-        choices=[season.value for season in Artist],
-        type=str,
-        help="The artist of the objekts",
+    artist_subparser = parser.add_subparsers(
+        dest="artist", help="Artist to run objekt-stats on."
     )
-    parser.add_argument(
-        "season",
-        choices=[season.value for season in Season],
-        type=str.capitalize,
-        help="The season of the objekts",
-    )
-    parser.add_argument(
-        "-c",
-        "--collection-no",
-        type=str.upper,
-        help="The collections to collect data for, e.g. 117Z,118Z,119Z,120Z",
-    )
-    parser.add_argument(
-        "--edition",
-        choices=[edition.value for edition in Edition],
-        type=Edition,
-        help="The edition to collect data for. There are three editions each season.",
-    )
-    parser.add_argument(
-        "-o",
-        "--output",
-        choices=[report.value for report in StatsOutput],
-        default=StatsOutput.TERM,
-        type=str.lower,
-        help="Outputs statistics in the provided format (defaults to term)",
-    )
+    for artist, season_enum in ARTIST_SUBPARSERS:
+        artist_parser = artist_subparser.add_parser(artist.value)
+        artist_parser.add_argument(
+            "season",
+            choices=[season.value for season in season_enum],
+            type=str.capitalize,
+            help="The season of the objekts",
+        )
+        artist_parser.add_argument(
+            "-c",
+            "--collection-no",
+            type=str.upper,
+            help="The collections to collect data for, e.g. 117Z,118Z,119Z,120Z",
+        )
+        artist_parser.add_argument(
+            "--edition",
+            choices=[edition.value for edition in Edition],
+            type=Edition,
+            help=(
+                "The edition to collect data for. There are three editions each season."
+            ),
+        )
+        artist_parser.add_argument(
+            "-o",
+            "--output",
+            choices=[report.value for report in StatsOutput],
+            default=StatsOutput.TERM,
+            type=str.lower,
+            help="Outputs statistics in the provided format (defaults to term)",
+        )
 
     # Show help text when no args provided
     if len(sys.argv) == 1:
