+++ src/objekt_stats/enums/cli.py
@@ -1,0 +1,8 @@
+from enum import StrEnum
+
+
+class StatsOutput(StrEnum):
+    TERM = "term"
+    CSV = "csv"
+    HTML = "html"
+    JSON = "json"
