+++ src/objekt_stats/objekts/models.py
@@ -1,10 +1,7 @@
 from pydantic import BaseModel
 
-from objekt_stats.enums import ArtmsMember, Season, TripleSMember
-
 
 class ObjektCollectionData(BaseModel):
-    season: Season
-    member: TripleSMember | ArtmsMember
+    member: str
     collection_no: str
     total: int
